server/requirements.txt
flask
pandas
Pillow
PyYAML
seaborn
thop
matplotlib
opencv_python
numpy
tqdm
scipy
coremltools
pafy
flask_cors
flask_jwt_extended
flask-swagger-ui
sqlalchemy
flask-bcrypt
youtube-dl
werkzeug
flask-restx
python-dotenv
flask_socketio
eventlet
flask_jwt_extended
flask_sqlalchemy
psycopg2-binary
#reqeuests
gevent-websocket
tensorboardX
termcolor
yacs
torchsummary
imageio
#pycocotools
vidgear
loguru

--find-links https://download.pytorch.org/whl/torch_stable.html

torch==1.9.0+cpu
torchvision==0.10.0+cpu
torchaudio==0.9.0
